Changes for level 4
Just docblock changes here.

I tried fixing as many of the is_float/is_numeric/is_string and
comparison errors as I could, but the rest would require code changes
to do correctly.

The is_null (and comparisons to null) are usually argument validation
to prevent values with `null` which is allowed by PHP unless there’s a
typehint disallowing it. I didn’t want to add null as an acceptable
type in the docblock in these situations.